Renamed temp_view to immediate_view
