+++ lib/immediate_view/bar.ts
