+++ lib/immediate_view/chartlayermanager.ts
