+++ lib/immediate_view/rule.ts
