+++ lib/immediate_view/scatter.ts
