+++ lib/immediate_view/serieslabelstrip.ts
