+++ lib/immediate_view/symbol.ts
