+++ lib/immediate_view/tickstrip.ts
